move common functions to common.hpp
